pyproject.toml
[project]
name = "datacollection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "openai-agents>=0.2.7",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
]
